+++ src/plugin/utils/props.ts
@@ -1,31 +1,32 @@
-import { Column } from '@/types';
+import { Column, DefaultColors } from '@/types';
+
+export const defaultColorValues: DefaultColors = {
+	background: '--v-theme-surface',
+	border: 'transparent',
+	color: '--v-theme-on-surface',
+};
 
 export const AllProps = {
-	// color: 'primary',															// ? Doesn't seem to work
-	colors: () => ({
-		default: {
-			background: 'primary',
-			base: 'primary',
-			border: 'primary',
-			color: 'on-primary',
-		},
+	// color: 'primary', // ? Currently only works with multiSort (colored number in header, not added to VDT) - Vuetify prop
+	colorPercentageChange: 15,
+	colorPercentageDirection: 'desc' as const,
+	colors: () => ({ 																	// ? Internal Prop
+		default: defaultColorValues,
 		footer: {
-			background: '--v-theme-surface',
-			color: '--v-theme-on-surface',
+			background: undefined,
+			color: undefined,
 		},
 		header: {
-			background: 'primary',
-			color: 'on-primary',
+			background: undefined,
+			color: undefined,
 		},
 		percentageChange: 15,
 		percentageDirection: 'desc',
-		table: {
-			bottomBorder: 'primary',
-		},
 	}) as const,
 	columnWidths: () => ([]),
 	// customFilter: undefined, 											// ? Needs Testing
 	// customKeyFilter: undefined,										// ? Needs Testing
+	defaultColors: () => defaultColorValues,
 	density: 'default',
 	// drilldown: () => ({}),													// * Works - Internal Prop
 	drilldownKey: 'id',
@@ -37,8 +38,12 @@ export const AllProps = {
 	// firstIcon: '',																	// * Works - Keep Commented Out
 	// firstPageLabel: '',														// ? Doesn't seem to work
 	// fixedFooter: true, 														// ? Doesn't seem to work
-	// fixedHeader: true, 														// ? Doesn't seem to work
+	// fixedHeader: true, 														// * Works - Keep Commented Out
+	footerBackgroundColor: undefined,
+	footerColor: undefined,
 	footers: () => [] as Column[],
+	headerBackgroundColor: undefined,
+	headerColor: undefined,
 	// headers: () => [] as const,										// * Works - Keep Commented Out
 	height: 'auto',
 	// hideDrilldownWhenLoading: true,								// * Works - Keep Commented Out
@@ -100,7 +105,7 @@ export const AllProps = {
 	loading: false,
 	loadingText: '$vuetify.dataIterator.loadingText',
 	matchColumnWidths: false,
-	multiSort: false,																	// ! Not sure if this is working correctly - Is binding prop
+	// multiSort: false,															// * Works - Keep Commented Out
 	// nextIcon: '', 																	// * Works - Keep Commented Out
 	noDataText: '$vuetify.noDataText',
 	noFilter: false,