+++ src/types/index.ts
@@ -58,13 +58,15 @@ export interface TableColors<T = any> {
 	[key: string]: T;
 }
 
+export type DefaultColors = {
+	background?: string;
+	base?: string;
+	border?: string | null;
+	color?: string;
+};
+
 export type ColorsObject = {
-	default?: {
-		base?: string;
-		background?: string;
-		border?: string | null;
-		color?: string;
-	};
+	default: DefaultColors;
 	footer?: {
 		background?: string;
 		color?: string;
@@ -75,15 +77,12 @@ export type ColorsObject = {
 	};
 	percentageChange?: number;
 	percentageDirection?: 'asc' | 'desc';
-	table?: {
-		bottomBorder?: string;
-	};
 };
 
 export type LevelColorResponse = {
 	background?: string;
 	base?: string;
-	bottomBorder?: string;
+	border?: string;
 	circular?: string;
 	color?: string;
 	linear?: string;
@@ -150,10 +149,13 @@ export type TableType = VDataTable | VDataTableServer | unknown;
 
 // -------------------------------------------------- Props //
 export interface Props {
-	colors?: ColorsObject | null;
+	colorPercentageChange?: ColorsObject['percentageChange'];
+	colorPercentageDirection?: ColorsObject['percentageDirection'];
+	colors?: NonNullable<ColorsObject>;
 	columnWidths?: number[];
 	customFilter?: VDataTable['$props']['customFilter'];
 	customKeyFilter?: VDataTable['$props']['customKeyFilter'];
+	defaultColors?: ColorsObject['default'];
 	density?: VDataTable['$options']['density'];
 	drilldown?: object;
 	drilldownKey?: string;
@@ -164,8 +166,12 @@ export interface Props {
 	filterMode?: VDataTable['$props']['filterMode'];
 	fixedFooter?: boolean;
 	fixedHeader?: boolean;
+	footerBackgroundColor?: string | undefined;
+	footerColor?: string | undefined;
 	footers?: Column[];
 	groupBy?: VDataTable['$props']['groupBy'];
+	headerBackgroundColor?: string | undefined;
+	headerColor?: string | undefined;
 	headers?: VDataTable['$props']['headers'];
 	height?: string | number | undefined;
 	// hideDefaultFooter?: boolean;																			// ? Custom Property - Need to add/test
@@ -267,7 +273,12 @@ export interface TopSlotProps extends VDataTableSlotProps {
 };
 
 export interface HeaderSlotProps extends AllSlotProps {
+	colorPercentageChange?: Props['colorPercentageChange'];
+	colorPercentageDirection?: Props['colorPercentageDirection'];
+	colors: Props['colors'];
 	columnWidths: Props['columnWidths'];
+	headerBackgroundColor?: Props['headerBackgroundColor'];
+	headerColor?: Props['headerColor'];
 	isTheadSlot?: boolean;
 	items: Props['items'];
 	loaderProps?: LoaderProps;
@@ -328,6 +339,10 @@ export interface ItemSlotProps extends Omit<AllSlotProps, 'colors' | 'sortBy'> {
 }
 
 export interface TFootSlotProps extends Omit<AllSlotProps, 'showSelect' | 'sortBy'> {
+	colorPercentageChange?: Props['colorPercentageChange'];
+	colorPercentageDirection?: Props['colorPercentageDirection'];
+	footerBackgroundColor?: Props['footerBackgroundColor'];
+	footerColor?: Props['footerColor'];
 	footers: Column[];
 	items: Props['items'];
 	selectStrategy: Props['selectStrategy'];
@@ -435,7 +450,7 @@ export interface UseGetHeaderColumnWidths {
 export interface UseGetLevelColors {
 	(
 		options: {
-			colors: ColorsObject | undefined | null,
+			colors: Props['colors'],
 			level: Props['level'],
 			prop: string,
 			themeColors: ThemeInstance,
@@ -447,7 +462,7 @@ export interface UseGetLevelColors {
 export interface ConvertLevelColors {
 	(
 		options: {
-			colors: ColorsObject,
+			colors: NonNullable<Props['colors']>,
 			level: Props['level'],
 			prop: string,
 			theme: ThemeInstance,
@@ -458,7 +473,7 @@ export interface ConvertLevelColors {
 
 export interface LevelPercentage {
 	(
-		colors: ColorsObject,
+		colors: Props['colors'],
 		level: number,
 		direction: string,
 	): number;
@@ -549,7 +564,6 @@ export interface UseCheckBoxClasses {
 export interface UseHeaderCellClasses {
 	(
 		options: {
-			colors: ColorsObject | undefined | null | false,
 			column: Column,
 			level: Props['level'],
 			slotName?: string,
@@ -621,9 +635,11 @@ export interface UseTFootRowClasses {
 export interface UseHeaderCellStyles {
 	(
 		options: {
-			colors: ColorsObject | undefined | null | false,
+			colors: Props['colors'],
 			column: { width?: string | number; },
 			dataTableExpand: boolean,
+			headerBackgroundColor?: Props['headerBackgroundColor'];
+			headerColor?: Props['headerColor'];
 			level: Props['level'],
 			theme: ThemeInstance,
 		}
@@ -633,7 +649,7 @@ export interface UseHeaderCellStyles {
 export interface UseTableStyles {
 	(
 		options: {
-			colors: ColorsObject | undefined | null | false,
+			colors: Props['colors'],
 			level: Props['level'],
 			theme: ThemeInstance,
 		}
@@ -643,8 +659,9 @@ export interface UseTableStyles {
 export interface UseTFootCellStyles {
 	(
 		options: {
-			colors: ColorsObject | undefined | null | false,
+			colors: Props['colors'],
 			elm: string,
+			footerColor?: Props['footerColor'],
 			level: Props['level'],
 			theme: ThemeInstance,
 		}
@@ -657,7 +674,7 @@ export type DrilldownEvent = {
 	$event?: MouseEvent | undefined;
 	columns?: object;
 	index?: number;
-	internalItem: DataTableItem | any;
+	internalItem: DataTableItem['internalItem'];
 	isExpanded: IsExpanded;
 	isRow?: boolean;
 	item: DataTableItem | any;
@@ -671,6 +688,7 @@ export type ClickRowCheckboxEvent = {
 	columns?: object;
 	index?: number;
 	item?: object;
+	internalItem?: DataTableItem['internalItem'];
 	level?: Props['level'];
 	toggleSelect(item?: object): void;
 };